backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# PostgreSQL extensions
pgvector==0.2.4

# Redis
redis==5.0.1

# Object Storage (S3/MinIO)
boto3==1.34.34
minio==7.2.3

# AI & ML
openai==1.10.0

# Web Scraping
patchright==1.55.2  # Playwright的修补版本，绕过CDP检测
beautifulsoup4==4.12.3
lxml==5.1.0
html5lib==1.1

# Data Processing
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
httpx==0.26.0
aiofiles==23.2.1
python-dateutil==2.8.2

# Monitoring & Logging
prometheus-client==0.19.0
